(chore) change settings type to JSONObject
