+++ imports/plugins/core/core/server/no-meteor/resolvers/index.js
@@ -24,11 +24,6 @@ export default {
     ping: () => "pong",
     ...Query
   },
-  PackageSettingsData: {
-    __resolveType(obj) {
-      return obj.gqlType;
-    }
-  },
   Shop,
   Tag,
   Time: GraphQLTime,
