+++ imports/plugins/core/core/server/no-meteor/schemas/index.js
@@ -4,7 +4,7 @@ import currency from "./currency.graphql";
 import email from "./email.graphql";
 import metafield from "./metafield.graphql";
 import node from "./node.graphql";
-import packageSettings from "./packageSettings.graphql";
+import pluginSettings from "./pluginSettings.graphql";
 import shop from "./shop.graphql";
 import tag from "./tag.graphql";
 
@@ -15,7 +15,7 @@ export default [
   email,
   metafield,
   node,
-  packageSettings,
+  pluginSettings,
   shop,
   tag
 ];
