--- imports/plugins/core/core/server/no-meteor/schemas/packageSettings.graphql
@@ -1,15 +1,0 @@
-type MailData {
-  user: String!
-  password: String!
-}
-
-type CoreSettingsData {
-  mail: MailData!
-}
-
-union PackageSettingsData = CoreSettingsData
-
-extend type Query {
-  "A generic query for getting the settings for a package. Requires admin role."
-  packageSettings(name: String, shopId: ID!): PackageSettingsData
-}
